订单的查询和创建

+++ backend/src/main/java/com/example/model/SearchRoomRequest.java
@@ -14,7 +14,7 @@ public class SearchRoomRequest {
     private RoomType roomType;      // 会议室类型
     private LocalDate date;         // 日期
     private Integer attendance;     // 参会人数
-    private Integer startTime;      // 开始时间，整数
-    private Integer endTime;        // 结束时间，整数
+    private Byte startTime;      // 开始时间，整数
+    private Byte endTime;        // 结束时间，整数
     private List<Long> requiredEquipments;        // 所需使用的设备的ID
 }
